// auctionhouseapp.client/src/dto/liveAuction/ILiveAuctionItem.d.ts
//     This code was generated by a Reinforced.Typings tool.
//     Changes to this file may cause incorrect behavior and will be lost if
//     the code is regenerated.

export interface ILiveAuctionItem
{
	itemId: string;
	name: string;
	description: string;
	image: string;
	startingPrice: number;
	reservePrice: number;
	status: string;
}